[Refactor] GetResellRes -> GetResellInfoRes 클래스 이름 변경

+++ src/main/java/site/lghtsg/api/resells/model/GetResellBoxRes.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class ResellBox {
-    private int resellIdx;
+    private long resellIdx;
     private String name;
     private String rateOfChange;
     private String rateCalDateDiff;
